+++ libs/voicefocus/types.d.ts
@@ -130,5 +130,4 @@ export interface VoiceFocusFetchBehavior {
 export interface VoiceFocusFetchConfig extends VoiceFocusFetchBehavior {
     paths: VoiceFocusPaths;
 }
-
 export {};