+++ libs/voicefocus/worklet-inline-node.js
@@ -1,3 +1,6 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 "use strict";
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
     function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
@@ -11,9 +14,11 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 Object.defineProperty(exports, "__esModule", { value: true });
 const support_js_1 = require("./support.js");
 const types_js_1 = require("./types.js");
+const CPU_WARNING_MAX_INTERVAL_MS = 5 * 1000;
 class VoiceFocusInlineNode extends types_js_1.VoiceFocusAudioWorkletNode {
     constructor(context, options) {
         super(context, options.processor, options);
+        this.cpuWarningCount = 0;
         this.channelCountMode = 'explicit';
         this.channelCount = 1;
         const { modelURL, worker, fetchBehavior, logger, delegate, } = options;
@@ -59,7 +64,15 @@ class VoiceFocusInlineNode extends types_js_1.VoiceFocusAudioWorkletNode {
         const data = event.data;
         switch (data.message) {
             case 'cpu':
-                (_a = this.logger) === null || _a === void 0 ? void 0 : _a.warn('CPU warning:', data.message);
+                this.cpuWarningCount++;
+                const now = Date.now();
+                const before = this.cpuWarningLastTriggered || now;
+                const diff = Math.abs(now - before);
+                if (!this.cpuWarningLastTriggered || diff > CPU_WARNING_MAX_INTERVAL_MS) {
+                    (_a = this.logger) === null || _a === void 0 ? void 0 : _a.warn(`CPU warning (count: ${this.cpuWarningCount}):`, data.message);
+                    this.cpuWarningCount = 0;
+                    this.cpuWarningLastTriggered = now;
+                }
                 (_b = this.delegate) === null || _b === void 0 ? void 0 : _b.onCPUWarning();
                 break;
             default: