+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [3.4.0] - 2022-05-24
 
 ### Added
+- Add the reserved status code AudioDisconnectAudio.
 
 ### Removed